code review comment resolved
