+++ api/helm-app/gRPC/applicationClient.go
@@ -28,7 +28,7 @@ type HelmAppClient interface {
 	IsReleaseInstalled(ctx context.Context, in *ReleaseIdentifier) (*BooleanResponse, error)
 	RollbackRelease(ctx context.Context, in *RollbackReleaseRequest) (*BooleanResponse, error)
 	TemplateChart(ctx context.Context, in *InstallReleaseRequest) (*TemplateChartResponse, error)
-	TemplateChartBulk(ctx context.Context, in []*InstallReleaseRequest) ([]*TemplateChartResponse, error)
+	TemplateChartBulk(ctx context.Context, in *BulkInstallReleaseRequest) ([]*TemplateChartResponse, error)
 	InstallReleaseWithCustomChart(ctx context.Context, in *HelmInstallCustomRequest) (*HelmInstallCustomResponse, error)
 	GetNotes(ctx context.Context, request *InstallReleaseRequest) (*ChartNotesResponse, error)
 	ValidateOCIRegistry(ctx context.Context, OCIRegistryRequest *RegistryCredential) (*OCIRegistryResponse, error)
@@ -294,13 +294,12 @@ func (impl *HelmAppClientImpl) TemplateChart(ctx context.Context, in *InstallRel
 	}
 	return response, nil
 }
-func (impl *HelmAppClientImpl) TemplateChartBulk(ctx context.Context, in []*InstallReleaseRequest) ([]*TemplateChartResponse, error) {
+func (impl *HelmAppClientImpl) TemplateChartBulk(ctx context.Context, in *BulkInstallReleaseRequest) ([]*TemplateChartResponse, error) {
 	applicationClient, err := impl.getApplicationClient()
 	if err != nil {
 		return nil, err
 	}
-	bulkReq := &BulkInstallReleaseRequest{InstallReleaseRequest: in}
-	response, err := applicationClient.TemplateChartBulk(ctx, bulkReq)
+	response, err := applicationClient.TemplateChartBulk(ctx, in)
 	if err != nil {
 		return nil, err
 	}
