+++ api/router/app/pipeline/configure/PipelineConfigRouter.go
@@ -57,7 +57,7 @@ func (router PipelineConfigRouterImpl) InitPipelineConfigRouter(configRouter *mu
 	configRouter.Path("/template/update").HandlerFunc(router.restHandler.UpdateAppOverride).Methods("POST")
 	configRouter.Path("/template/list").Queries("appId", "{appId}").Queries("envId", "{envId}").HandlerFunc(router.restHandler.GetTemplateComparisonMetadata).Methods("GET")
 	configRouter.Path("/template/data").HandlerFunc(router.restHandler.GetDeploymentTemplateData).Methods("POST")
-	configRouter.Path("/template/rotate-pod").HandlerFunc(router.restHandler.GetRestartWorkloadData).Queries("appIds", "{appIds}").
+	configRouter.Path("/template/workloads").HandlerFunc(router.restHandler.GetRestartWorkloadData).Queries("appIds", "{appIds}").
 		Queries("envId", "{envId}").Methods("GET")
 	// Start -- user defined gitops repository
 	configRouter.Path("/template/gitops/config/{appId}").HandlerFunc(router.restHandler.GetGitOpsConfiguration).Methods("GET")
